Bump wlroots build in CI to 0.14.0
wlroots released 0.14.0 recently, and pywlroots has released
accordingly. We should bump up wlroots' version in the CI (pywlroots
always just gets the latest via pip).

The new version has support for running wlroots headlessly, meaning that
it should run as expected in GitHub actions so we can start putting
together tests for the Wayland backend. One downside to this is that
some of the dependencies need to be built manually as Ubuntu's repos are
very outdated. For now, adding more manual builds to the GA setup works
fine, but if in the future we foresee this list growing we could
consider a different approach such as running an up-to-date distro in a
docker container and installing these dependencies from their repos.