WL: Map Button1-9 codes to correct mouse events
The current button codes used for binding button events to
`config.Mouse` "ButtonX" strings are inconsistent with X11 and by
extension what people expect. This aligns how we define buttons 1-9 with
how Sway does this, which is consistent with how X11 does it.

Buttons 1-3 and 8-9 are true "buttons" whereas buttons 4-7 map to scroll
events (up, down, left, right), which are received via cursor axis
events and are then converted to buttons and processed by us as buttons
in case there are things bound to them. As the scroll "buttons" are
mimicked by us, their values are defined as above `KEY_MAX` so that no
button events that we receive can trigger these events.