Use Gtk based window for tests for X11 and Wayland support
This replaces `test/scripts/window.py`, which creates a simple test
window with xcffib, with a script that creates a simple window with Gtk.
This lets us use the same script to create the same basic window under
both X11 and Wayland.

This is used in tests with `manager.create_window()` as normal. However
the signature of this has slightly changed: `name` can still be passed
to specify a title, but now an optional boolean `floating` arg can be
used to change the wm_class (X11) / app id (Wayland) to the string
"dialog", which is used by the default floating rules to float the
window. This replaces `manager.create_dialog` but is more explicit in
its purpose.

Some tests want to create a window that does not receive focus. These
are almost exclusively notifications, so `manager.test_notification` can
still be used, telling the same script to set the window to be a
notification. Currently this does not work under Wayland due to no
wlr_layer_shell support in Gtk.

Lastly `test_window.py::test_no_size_hint` has some values changed due
to the new test window having default desired dimensions used when
floating.