Fix 'qtile run-cmd' bug with no command
Executing `qtile run-cmd` with no arguments results in an ugly
exception with no explanation of the problem. This is because
the script tries to execute the `cmd` argument and errors when
there's no command.

The problem is that the script uses `argparse.REMAINDER` to catch
the command and appropriate arguments but `REMAINDER` allows no
matches whereas we need at least one.