Let tests be run on one or both backends
This is the first of some commits that will let the test suite be run on
either one chosen backend or both backends.

Some variables and classes (e.g. `WIDTH`, `Retry`) are moved into
helpers.py to avoid circular imports between test/conftest.py and
test/backend/<backed>/conftest.py.

Most x11-specific helpers in test/conftest.py are moved into
test/backend/x11/conftest.py. The exception to this is `xephyr` as this
provides a session-scoped fixture required at all levels of the
directory structure.

Both the x11 and wayland backends have an environment-setting context
manager defined in their conftest.py that is used to set a
session-scoped environment. Eventually this will/should handle the
xephyr stuff too. These would then be passed backend-agnostic settings
such as the number of outputs to use (a skeleton `outputs` fixture is
added which will be extended and will replace `xephyr` where it gets
parametrized by some tests).

A basic `Backend` class is defined in conftest.py to handle passing
arguments to the current `Core` class and carrying the dict of
environmental variables that should be passed to an subprocesses during
tests (i.e. `DISPLAY`, `WAYLAND_DISPLAY`).

Lastly, the backends to use for tests can be set on the command line:

    python -m pytest test --backend=x11
    python -m pytest test --backend=x11 --backend=wayland

By default, only x11 is tested.