Ensure tests and the two test processes have the right env vars
This changes how the backends are passed to TestManager and set up for
tests. The main requirement that this addresses is that the
environmental variables of the child process and parent must both be
correct. When the Wayland core is instantiated in the child process it
finds the next free display name and sets that in `os.environ`, and so
the parent process must get this information and set it when creating
test windows. In contrast the x11 Core is given a pre-determined DISPLAY
to use, which is easier to handle - it's just passed along in the
`Process` call.