Add actions and NotificationClosed method to notifier and notify widget
This adds action handling to the notifier, which uses the
`ActionInvoked` signal on the notification service. This is bound to
button3 on the Notify widget. This will do things such as take Firefox
to the tab that sent a notification.

This also fixes the `NotificationClosed` signal which needs to return
the values it is passed to work. This signal is emitted when a
notification is closed and tells listeners of the reason, which can be
one of: expired (due to timeout), dismissed (the user cleared it) or
method (it was cleared by the dbus client via the CloseNotification
method). The CloseNotification method is given a basic implementation
whereby code registered with the notifier can pass an on_close callback
to `notifier.register` which will get executed when a dbus client calls
this method.