WL: fix cmd_info on Window classes
Currently cmd_info was returning nothing when called via the client
interface. It appears to always use the `info` method, so `cmd_info`
can't have a proper implementation and so is redundant. Only `info`
needs to be defined on a subclass of the base `Window`.

Also, there is nothing additional to the `info` data that `cmd_inspect`
would return in the Wayland backend, so rather than
implementing/proxying it, let's not at all, and move the test for
`cmd_inspect` into test/backend/x11.