Fix bug in test_incompatible_widget due to mutable config class
When running this test a second time (i.e. when testing both x11 and
wayland) the screens of the test `Config` class are `_configure`-ed a
second time, which leads to problems with `qtile.core.masked` in the
in-test-thread Qtile test instance. Instead the `Config` should be fresh
each time.