Let test_fakescreen work with both backends
This changes the fakescreen tests to use the same output dimensions of
the other tests. This simplifies setup of the tests as the width and
height of outputs do not need to be configurable. This makes these tests
work with the Wayland backend as the test output dimensions now match.

This also revealed a bug in the Wayland Window class where moving the
window to a different screen did not cause the window to change to that
group, when it should.