WL: create window ID when being managed, not created
Windows become managed by `Qtile` when they are first mapped, and not
necessarily when they are first created - when they wait in a list of
pending windows. Their window ID should be created when mapped and
becoming managed, rather than being added to the pending list, otherwise
two windows could be given the same ID as the ID is determined using the
maximum ID found in `qtile.windows_map`.