WL: Fix floating_only bring_front_click behaviour
This fixes the "floating_only" behaviour of the `bring_front_click`
config option in Wayland which was not set up properly. The Wayland test
fake_click is changes slightly so that the pointer is warped before
clicking so that the correct window is used for restacking, and pointer
buttons are not processed. If this behaviour is needed for a future test
then it can be implemented then. Currently it is not, as internal code
will need updating so that the core can read keyboard modifier state
without a keyboard attached; the test wayland environment does not have
a functional keyboard.