Allow keyboard layout widget to specify variant under Wayland
When running setxkbmap under the X11 backend, the keyboard variant is
passed as the argument to the binary after the layout name, and so is
handled by splitting the layout name on the space and passing these
along to the binary.  For the `keymap_new_from_names` call, the variant
needs to be explicitly passed in as a separate argument.  With this
change, perform the equivalent operation for the layout name string
passed to the wayland backend of the keyboard layout widget to parse out
the variant when it is specified and forward it through to the xkbcommon
call.