Add dependency tables to docs
This adds two tables to the docs that list out in one place (well, two
places) the dependencies needed for 1. general runtime usage and 2.  the
tests. These have replaces some more verbose explanation of how to
install individual things, the assumptions being that people know how to
install from pip, and people know how to find packages for their own
distro if they know the name of the package for Ubuntu.

The first table is on the installation page and lists core dependencies,
whereas the table listing tests dependencies is on the hacking page and
lists additional dependencies.

What is *not* included is widget dependencies. There are many, all are
optional, and they change more quickly than the rest of the deps, so
splitting where they are documented between two places is probably a bad
idea. However, many of the dependencies are monkeypatched for testing
and so fortunately most are not needed for tests. Those are are needed
for tests are in the test deps table.

Closes #2466.