WL: only deactivate previously focussed xdg_surface if it isn't next
Determining when to deactivate the previously focussed xdg_surface is
based on the seat's focussed wlr_surface, so if the next and current
wlr_surfaces differ but belong to the same xdg_surface we end up with a
focussed window with a deactivated xdg_surface. This results in the
'unfocussed look' used by GTK programs, for example, even when focussed.