Fix systray bug where qtile is restarted
This commit changes the timing of setting the Systray's visual id.
It needs to be set earlier so, if apps are already running when the
sytray is restarted, the visual id needs to be set before the tray
announces itself as icons will try to reigster at that point.

Fixes #2646