x11: Do not set straight-to-Static window geometry to 0
If a newly created window gets `cmd_static`-ed right away (due to having
the dock window type), it might not have its width and height set yet,
and so we configure it with placeholder values of 0. This results in an
X protocol value error. Instead we should fall back to non-zero values.
I've put 200 for width and height which seems reasonable. The window
should then set its own geometry after mapping in that case.

Also - static windows created in this way need to unhide themselves, so
that's added too.