WL: set seat keyboard upon modifier press
When keyboard devices receive modifier events they should set themselves
as the seat's keyboard in case they are not. This prevents the
occasional bug where a true (typing) keyboard is not the seat's current
keyboard (another device with at least 1 button might be) and so its
modifiers are not used for keybindings.