Allow KeyChord mode names without staying in mode
This PR allows all KeyChords to be named (via `mode`)
but gives the user the option to switch off the mode
persistence then the key sequence is complete.

The benefit of this is that the mode name can be displayed in
the Chord widget for all KeyChords, not just those with
persistent modes.

Closes #2647