Group: Make sure both tiling and floating windows are added to the layout
Windows are added to the list and are not removed when they are set to
floating. Instead the function float_client is called. When a client is set to
tiling, tile_client is called.
When a window is actually removed from the group, we need to make sure
to remove it regardless if it was floating or not.

Co-authored-by: whatacold <whatacold@gmail.com>