Layout Base: Restructure _ClientList to only expose tiled clients
_ClientList now only exposes tiled clients (clients that are not floating) to the
user. To get the tiled clients, the property clients needs to be
called. For most layouts, this is also handled in _SimpleLayoutBase as
that use _ClientList.

The Layout class gets the functions tile_client and float_client that
can be implemented by layouts, however, they should not be needed in
most cases. These functions are only called in the group code.

Co-authored-by: whatacold <whatacold@gmail.com>