Layout BSP: Handle cases where a node is floating in _BspNode
As the BSP layout uses a different data structure than a
list (_ClientList), we need to make sure to handle the case where
a node is set to floating.
This involves setting the proper split ratio, distance and setting the
shortest path to infinity so that these nodes are never picked.

Co-authored-by: whatacold <whatacold@gmail.com>