Using decorators to expose commands
Currently, commands that wish to be exposed via IPC need to be
prefixed with `cmd_` but this prefix is then removed when calling
commands. This can cause confusion for users.

This PR replaces the prefix mechanism with a new decorator:
`@expose_command` (imported by `from libatile.command.client import
expose_command`). Methods decorated with this will be exposed
via IPC but the method name is is now consistent in both IPC and
internal calls.

Refer to CHANGELOG for more details.