Add initial basic handling of Z axis
This is a reimplementation of #2194 by @tcld, itself a reimplementation
of #2029 based on the work of @senthilbaboo in #2144. The central
function `Window.change_layer()` is only provided in a rudimentary form,
so the current behaviour remains unchanged.

This also adds a Core method, `update_client_stack`, which the qtile
manager uses to tell the backend to update its representation of mapped
windows to reflect any changes in Z positions.

Closes #753.