x11: toggle window button grabbing with focussing
This changes how the x11 backend handles grabbing of button events on
windows. Currently, button events are grabbed for button 1 when new
windows are created so that we can listen to these and focus an
unfocussed window upon click. This works for focussing, but the grabbing
of buttons even on the focussed window can cause some problems, such as
that seen in #2744. This new approach changes this so that the button
events are grabbed only when a window is unfocussed, and ungrabs them
when the window gains focus. This follows the approach used by i3, and
also reduces some unnecessary button event processing as we no longer
respond to button events on the focussed window, whose button events we
do not care about.

Fixes #2744