Layout Base: Pseudo add and remove toggled floating clients
This makes sure so that the current index is set properly after a
client is set to tiled/floating. This is important for e.g. shuffling
clients.

It is implemented by giving a pseudo boolean to add and remove that is
set to True if the client list should not be altered.