WL: handle output scaling
This lets the wayland backend correctly handle scaling of outputs. Some
things were already partially handled, such as window placement and
geometries. This extends the handling so that internal windows and
native wallpapers are positioned and scaled correctly, and so that the
correct output region is rendered to, making output scaling actually
usable.

The remaining issue with scaling appears to be some off-by-one errors
due to rounding of floats to ints that can manifest as a single pixel
gap between windows and borders, and possible a single-pixel outline at
the left and bottom edges of an output where rendering is not quite
hitting all of the available pixels.