Fix memory leak with widgetbox
b2b1ed8 changed the way that surfaces are drawn. When the widgetbox
hides widgets, it replaces `self.drawer.draw` with a no-op. However,
this prevents the drawer from clearing its `RecordingSurface`
object. As a result, the RecordingSurface will continue to accumulate
details of all draws made by the widget while it is hidden.

This PR clears the RecordingSurface while widgets are hidden.