WL: Implement window configure-ack-configure dance
`Core.masked` is a context manager that lets us apply a series of
changes to windows under a set of temporary conditions, for example to
ignore mouse-enter events when changing group in the case of the x11
backend. This commit uses this mechanism for a similar reason for the
wayland backend: to perform a series of operations which we can gather
into pending state changes. We then listen to the ack-configure event on
windows to wait until all windows that were involved in that state
change have acked. When the last one acks (or when 200ms have passed,
whichever comes first), we can then apply all of those state changes
simultaneously.