Refactor management of window geometry
This commit does a few things with the goal of having `Window.place()`
as the one and *only* way to change a window's x, y, width or height.
This lets the backend have a greater degree of control over window
reconfiguring, and is required for truly atomic state changes in the
Wayland backend, as well as correctly sending configure notifys to
windows in the X11 backend.

Changes:

- All arguments to `Window.place` are optional keyword args.
- The X11 `Window.place` correctly sends configure notify events when
  the window is moved but not resized.
- The Wayland `Window.place` only sends a resize request to the client
  when either `width` or `height` is passed.
- `Window._enablefloating` and `Window._reconfigure_floating`, which
  exist in both backends, have been collapsed into one (the latter).
- All instances where `Window.{x,y,width,height}` were set directly have
  been adjusted to instead pass the new value into `Window.place()`.
- All instances where `bordercolor=None` was passed to `Window.place`
  have been removed.
- All instances where the arguments passed to `Window.place` were
  directly retrieved from that window's property have been removed (See
  e.g. the change in scratchpad.py).

The intended usage is now to do something like `place(x=x, y=y)` or
`place(width=width, height=height)` if these are the only changes,
rather than directly setting those properties. Note that though these
properties should eventually become read-only by removing the setters,
the setters remain usable, logging a deprecation notice.