WL: Don't listen to toplevel events until first mapped
We aren't interested in events from windows that are not yet managed.
This can cause problems such as firing the `client_name_updated` hook
with a window that hasn't got a WID yet. We should just get this
information when it gets mapped and managed.

Similarly, the destroy event can be fired before we get mapped for the
first time, which raises an exception due to accessing `self.wid` before
the window is given a WID, as Qtile tries to unmanage it before it ever
gets managed.