WL: Atomically apply border changes during configure sequence
If a window is being resized or moved, geometry changes are added to the
queue of pending changes so that they can all be applied atomically.
This commit includes border width and colors in this process, if they
were provided to `Window.place()`.

This prevents border changes happening out of step with other layout
changes, which results in a strange state for a small number of frames
while waiting for windows to ack, where a window's border has changed
(possibly causing it to shift diagonally) but the remaining layout
changes are still pending.