WL: Add wlr_pointer_constrains so windows can constrain pointer motion
This uses wlroots' interface for using the
pointer-constraints-unstable-v1 Wayland protocol to allow clients to
request that the pointer be constrained to a specified region. This is
used for example by games and other programs that use pointer input for
control but don't want to the pointer to leave their window (or a
sub-region within their window).