WL: add support for relative-pointer-unstable-v1 protocol
This uses wlroots' interface for supporting the
relative-pointer-unstable-v1 Wayland protocol. From the protocol: "This
protocol specifies a set of interfaces used for making clients able to
receive relative pointer events not obstructed by barriers (such as the
monitor edge or other pointer barriers)." This is useful for example by
games and other programs that may request pointer contraints but want
to continue to receive pointer input events.

Requires pywlroots 0.14.6