Try to get version from importlib.metadata
This makes the `main` script try to find the installed `qtile` version
from `importlib.metadata`, which is available from Python 3.8 onwards.
This removes the `pkg_resources` dependency. `pkg_resources` is still
used as a fallback when used with Python 3.7.

Also, the try block that does this falling-back should be catching
`ModuleNotFoundError` which was not being caught in the case that
`pkg_resources` was not installed.

Fixes #2809