Fix rendering wallpapers on X11
Currently, when the root window dimensions change (e.g. after
changing the position of a monitor from side-by-side to
above-below), the pixmap for the toot window is not recreated.
The result is that images cannot be rendered to the new area.
This commit addresses the issue by checking dimensions before
updating the wallpaper.

Closes #2836