Add ability to pass parameters to lazy calls
Where users create functions in their config which are
decorated with `@lazy.function`, there is no easy way to
allow arguments to be passed to the decorated function when
attaching to a keybinding.

One alternative is to use a closure to wrap the decorated function
but this is more complicated for users.

The alternative, which is presented by this PR, is to create a
`__call__` method in the `LazyCall` class so that, when a user binds
a key in their config by doing `Key(..., decorated_function(args))`
those arguments are added to the `LazyCall` object and passed to the
decorated function when the key is pressed.

See https://github.com/qtile/qtile/issues/2848#issuecomment-934093883
for more detail on this.