Use base Window class for documenting window commands
This makes the docs point to `libqtile/backend.py::Window` to getting
information on window commands. Many of the changes synchronise the
command methods between the three `Window` classes, making sure the
wayland implementation has all of the commands and moving some methods
to the base class to be shared across them.