Better handling of incompatible widgets
The Systray widget crashes on Wayland as it is not compatible with
that backend. Users often report this as if it is a bug.

This PR adds better handling of widgets that may be incompatible
with certain backends. It does this by adding a
`supported_backends` attribute. The default value, an empty list,
means a widget is compatible with all backends. Where there is
limited compatibility then the compatible backends are named in
the list.

Documentation is updated to show backend compatibility by exception
only (i.e. users are only informed about compatibility of widgets
where a specific widget is not available on all backends).

Incompatible widgets are removed from the bar during `_configure`
however they are no longer replaced with a `ConfigError` widget.
Instead, the widget is simply removed and a message is included
in the logs.