WL: Set tiled edge xdg_surface state when mapping
We currently tell a window's xdg_surface to tile its edges immediately
when it is created. This causes problems with some clients (e.g.
signal-desktop) which do need appear to be ready for being configured.
Instead let's delay setting this state until the client is ready and
emits the map event. This is also what sway does.

Fixes #2835