Fix scratchpad focus issue in X11
There is an issue with some scratchpads where the window will
alternate opening with and without focus which can cause issues
for certain keybindings.

The issue seems to be that `set_input_focus` tries to focus the
root window rather than the scratchpad window.

This PR prevents focussing of the root window. This may or may not
cause other issues so needs testing!

Fixes #2588