Persist ScratchPad groups across config reloads
As part of the config reload functionality, we decided to not persist
groups across reloads as they do across restarts. This was because it
becomes difficult to consolidate existing groups that may or may not be
reflected in the config with new groups added to the config and groups
removed from the config.

However, this also applied to scratchpads, which can not be dynamically
changed during normal runtime and so do not pose the same issues. The
effect of doing this is that following a config reload each
scratchpad groups is fresh and old scratchpad windows get orphaned to
the current group. Instead, let's persist scratchpad groups across
reloads to prevent this and prevent this unexpected behaviour.

Fixes #1824.