WL: Fix no Internal clicks when bring_front_click is floating_only
This error is caused by `win.floating` raising an exception because
`Internal` doesn't have a float state. This commit re-writes the
bring-to-fron conditions to be more easy to read and explicitly check
that the window is not internal. It now passes this exception just fine
for all 3 window classes. As a result, the code continues and forwards
clicks to internal windows.

Closes #2793