Split space equally between consecutive stretch widgest
Currently widgets are grouped into blocks by splitting the list of
widgets on stretch widgets, and the position of each of these blocks on
the bar is centred such that the the blocks are distributed equally
across the bar.

If you have 2 consecutive stretch widgets configured, this can cause
expected effects where between them you have an empty block containing 0
widgets. This is amongst the blocks distributed equally, which can make
other blocks appear shifted and is confusing. Consider:

```py
widgets = [Left(), WindowName(), Spacer(), Centred(), Spacer(), Right()]
```

Here one might expect the `Centred()` widget to be centred but it will
not, because `WindowName` is a stretch widget, there are actually four
blocks, where the second between `WindowName` and `Spacer` is empty.
This puts `Centred` at 2/3 rather tahn 1/2 across the bar.

This is hard to track down between `Spacer` is invisible and looks like
part of `WindowName`, and the boundary is unclear causing issues with
mouse clicks going to the wrong widget.

This commit makes any number of consecutive stretch widgets be
considered together for the positioning of these blocks and makes them
split the space betwee blocks amongst them. In the example above this
means that positioning is calculated is if there was no first `Spacer`,
and then after that is done, it gets half of `WindowName`s length and
`WindowName` gets only half, correctly centreing `Centred` as would be
expected.

While the obvious question is "what if a user wants two adjacent stretch
widgets", the rationale for implementing this solution (and forcing such
users to stick a 0-width widget between those adjacent stretch widgets)
is that there isn't really a use-case for doing so - how would you
distribute space among them?

Fixes #2487.