Add test for undocumented commands
To ensure that our docs are as helpful as possible, all exposed
commands (e.g. those available via `lazy` etc.) should have a
docstring.

This script will be run as part of the tox suite and will list any
undocumented commands.

Currently, the script will not cause CI to fail (as there are a lot
of undocumented commands). Once this is fixed, the `-w` flag should
be removed so any new, undocumented commands will result in a
failure.