Ensure all widgets added to widgets_map
When finalising widgets, we loop over every item in `qtile.widgets_map`.
However, as this is a dictionary with the widget's name as the key,
using multiple instances of a widget means that some widgets are not
mapped as the key is duplicated.

This PR addresses the issue by ensuring that widgets are given a unique
key in `qtile.widgets_map`. One consequence of this is that widgets
will not be accessible via `lazy.widget[name]` unless the new key is
used. Users wishing to access widgets in this scenario should ensure
they rename widgets in their config.