Add FAQ section on sticking groups to screens
This is a frequently asked question so fits perfectly 😄

The implementations here are minimal but can serve as a starting points
if users want more specific behaviour (e.g. if you want to use
`toscreen(toggle=True)` then it does get a bit more complicated.)

Similarly this demonstrates some of what users can customise so if they
want to write a binding that scrolls between groups bound to a screen
they have a starting point of how to go about doing that.