Add Python 3.10 tox env
This adds Python 3.10 to tox and github actions, and runs the non-test
tox envs (mypy, lint etc) with Python 3.10.

This change only adds the new env, and does not remove py37. For the
time being it might be worth supporting py37 while we can, along with
pypy which is only compatible with python 3.7.