Grab keys in async mode when we swallow
Clients such as dmenu immediately grab the keyboard on opening. Dmenu
is used for extensions such as CommandSet.
This is a problem as when we grab the key in sync mode, it is only
ungrabbed on key release. Grabbing the key in sync mode is only needed
when we want the key to not be swallowed (passed to the client), so
let's grab in Async mode when it should be swallowed.

Co-authored-by: m-col <mcol@posteo.net>