WL: Fix unfocussable static windows
There is a bug where focussing static windows fails because
`win.ftm_handle` for the static window is accessed in
`Core.focus_window` but this attribute doesn't exist on static windows.
This change sets the attribute on static windows and uses it to enable
close requests via the protocol. Other requests don't make sense for
static windows (i.e.  maximise, minimise etc) so are not listened for.