X11: only warp cursor when focusing a new window
There is a bug in X11 where an active window sends a focus event
(e.g. opening a link in a new tab) which results in the cursor
being warped if `cursor_warp` is set to `True` in the config.

This PR addresses that by only warping when the focus event applies
to a new window.

Fixes #3135