Clicking on bar should focus that screen
Currently, clicking on a widget on a different bar does not focus that
screen. This results in undesirable consequences such as changing group
but new windows open on focused screen (#3165) and changing layouts on
focused screen (#3166).

This PR addresses that issue by checking the whether the bar's screen
is the focused screen when the bar is clicked. If the bar is not on
the focused screen then the screen is focused before the click event
is passed to the widget.

Fixes #3165
Fixes #3166