WL: Don't AttributeError during Static.get_wm_class()
When this method is called and the static window is a layer-shell
surface, the window has no `_app_id` attribute set. This needs to be set
manually in `Static.__init__`, just to `None`, so that when
`get_wm_class()` is called and `self._app_id` is checked, it doesn't
AttributeError.