Fix finalisation of layouts
Previously, when finalising layouts, the manager loops over
`self.config.layouts`. However, layouts are cloned per group and so
the config layouts do not represent the layouts that are actually
in use.

This is particularly problematic for the `TreeTab` layout which creates
an `Internal` window for the panel which needs to be killed when the
layout is finalised.

This PR fixes this issue by looping over layouts in each group to
finalise them. It also ensures that the `TreeTab._panel` is killed.

Fixes #3171