x11: only update wm_class upon property notify event
Currently the wm_class of a window is fetched with a get_property call
every time `window.get_wm_class` is called. In addition to unnecessary
calls, this can cause problems if `window.get_wm_class` is used in an
IPC call. Instead, let's save the wm_class on an attribute and only
update it when the window informs us of a change via a property_notify
event.