WL: Implement XWayland support
This is adds support for X-only clients through XWayland.

XWayland must be installed on the system. It's lazy to startup, so will
only be started when the first X connection is made. This adds a bit of
delay to that first client's startup, but it's better there than at
Qtile's startup. Users could start it up themselves in a startup hook to
avoid the delay.

Override-redirect windows (such as rofi, dmenu) are managed as `Static`
windows, which seems to work fine. Everything else is a `Window`.

Support is provided by wlroots, via pywlroots, which must have been
built with XWayland support (i.e. if XCB was present on the system when
it was build).