Prevent multiple instances of Systray
The System Tray specification does not appear to support multiple
trays. This can lead to issues where multiple trays are created.

This PR addresses the issue by causing Systray to raise a ConfigError
if multiple instances are created, or if a user attempts to create a
mirror of an existing Systray instance.

Fixes #2336