Fix more issues with Systray in WidgetBox
WidgetBox tries to reconfigure widgets when it is, itself,
reconfigured.

There is a bug where widgets in the box weren't being marked
as configured since we moved the setting of this flag
to Bar._configure.

This is particularly problematic when Systray is in the box for
two reasons:
1. If the box is open when the bar is reconfigured, the bar sees
an unconfigured Systray and tries to configure it and so creates
a config error.
2. If the box is closed, widgetbox still tries to reconfigure the
Systray and so also triggers a config error.

This PR addresses both of these scenarios by:
1. Ensuring the configured flag is set by the widgetbox
2. Only configuring widgets in the box when the box is configured
for the first time.
3. Adding tests for systray and widgetbox for `screens_reconfigured`
events.