Fix hard-coded ratio in Monad layouts
The ratio after using `layout.MonadTall.cmd_reset` was hard-coded to
0.5. Now it uses the ratio provided in layout's configuration, or
in the optional keyword argument (set it to 0.5 for old behavior).

The `_maximize_main` threshold is also affected - master will be
minimized if it's larger than the ratio configured by the user, not the
hard-coded 0.5.

The main pane ratio in Monad layouts can now be directly set to a
specified value.

The changes also affect MonadWide and MonadThreeCol because of inheritance.

Fixes #3270