x11: Only relayout upon hint update when increments update when floating
When `WM_NORMAL_HINTS` are updated we are uncondtionally re-laying out
the group. The client 'reaper' is getting into a never ending loop
because when it is placed it updates its hints. The `group.layout_all()`
call was added in 07e658a so that updates to width and height increment
hints are respect. This change removes the call except for specifically
when these hints are updated while the window is floating, as we do not
respect these hints in tiled windows anyway.

Fixes #3292