Xwayland: ignore startup geometries of 1 or below and fetch later
For some reason xterm starts up with a width and height set to 1. We
take startup width and height and save them as float with and float
height. Normally, if these are 0 (i.e. unchanged from __init__), when
the window is made floating the values are fetched from the client's
state again. Because xterm sets these to 1, we make them non-zero and do
not fetch them when it is made floating. This makes the windows tiny
when made floating for the first time which makes no sense.