Fix broken test_focus_change_event
This test is not doing what we think it does. The parametrized configs
are not loaded by the `Qtile` instance because `self` is being passed to
`super().__init__` and taking the `file_path` kwarg, which raises an
exception in `Config.load()`. This change fixes this and adds an
assertion to ensure that the right config is in fact loading.