Fix reconfigure screens bug
When reconfiguring screens, we checked to see if the screen size or
position has changed. If they had, we reset the `_configured` flag
on the bar to reconfigure it again. This had an unintended side
effect where we had configured widgets in an "unconfigured" bar.
In that scenario, the bar replaces the widgets with `Mirror`
instances which causes a number of issues (including `Systray`
raising a `ConfigError`).

This PR fixes the issue by leaving the `_configured` flag unchanged
but passes a `reconfigure` argument to the bar to force it to
reconfigure its size and position when required.