Fix reload_config for split configs
Using `reload_config` when a user has split their config over
multiple files results in the bar disappearing.

This can be fixed by reloading the modules that are imported by
the config.

This PR adds a loop to `reload_config` to reload any module that
is inside the same folder as config.py.

Fixes #3377