WL: output power management: only unblank outputs that blanked
Currently the wayland core indiscriminately blanks and unblanks outputs
(via the output power management protocol) without considering which
should be enabled or disabled. Outputs may be disabled through user
configuration of the output layout. We should only unblank outputs if it
was the power management protocol that blanked them in the first place.